requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
pyyaml==6.0.1
python-dotenv==1.0.1
httpx==0.27.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Code quality
black==24.1.1
ruff==0.2.1
mypy==1.8.0
